feat: 无符号类型处理，避免真实情况出现异常

+++ charge-pile-machine-common/src/main/java/com/huamar/charge/common/protocol/DataPacket.java
@@ -40,17 +40,17 @@ public class DataPacket extends BasePacket {
 	private byte msgBodyAttr;
 
 	/**
-	 * 4.消息体长度（2）
+	 * 4.消息体长度（2）原始 short 无符号
 	 */
-	private short msgBodyLen;
+	private int msgBodyLen;
 
 	/**
 	 * <pre>
 	 * 5.消息流水号（2）
-	 * 从１开始，每包消息加１，当达到 65535 后，重新从 1 开始
+	 * 从１开始，每包消息加１，当达到 65535 后，重新从 1 开始 原始 short 无符号
 	 * </pre>
 	 */
-	private short msgNumber;
+	private int msgNumber;
 
 	/**
 	 * 6.设备识别码（18）
