+++ charge-pile-machine-server/src/main/java/com/huamar/charge/pile/entity/dto/event/PileChargeArgConfigDTO.java
@@ -17,16 +17,16 @@ import lombok.EqualsAndHashCode;
 public class PileChargeArgConfigDTO extends PileEventBaseDTO {
 	
 	// 单体动力蓄电池最高允许充电电压 WORD 单位：0.001V/bit， 偏移 0V
-	public short maxValidVoltage;
+	public int maxValidVoltage;
 
 	// 最高允许充电电流 WORD 单位：0.1A/bit， 偏移-1600A
-	public short maxValidElectricity;
+	public int maxValidElectricity;
 
 	// 动力蓄电池标称总能量 D WORD 单位：0.1KW.h/bit，偏移 0
 	public int maxTotalPower;
 
 	// 最高允许充电总电压 WORD 单位：0.1V/bit，偏移 0
-	public short maxValidTotalVoltage;
+	public int maxValidTotalVoltage;
 
 	// 最高允许温度 BYTE 单位：1℃/bit，偏移-40℃
 	public byte maxValidTemperature;
@@ -35,20 +35,20 @@ public class PileChargeArgConfigDTO extends PileEventBaseDTO {
 	public byte carElectricityState;
 
 	// 整车动力蓄电池总电压 WORD 单位：0.1V/bit，偏移 0V
-	public short carTotalVoltage;
+	public int carTotalVoltage;
 
 	// 充电机同步时间 BCD[6] 见表 14，年月日时分秒
 	@JSONField(serializeUsing = BCDHexValueSerializer.class)
 	public BCD synTime;
 
 	// 充电机最高输出电压 WORD 单位：0.1V/bit，偏移 0V
-	public short chargerMaxOutVoltage;
+	public int chargerMaxOutVoltage;
 
 	// 充电机最低输出电压 WORD 单位：0.1V/bit，偏移 0V
-	public short chargerMinOutVoltage;
+	public int chargerMinOutVoltage;
 
 	// 充电机最大输出电流 WORD 单位 0.1A/bit，偏移-1600A
-	public short chargerMaxOutElectricity;
+	public int chargerMaxOutElectricity;
 
 	// BMS 充电准备就绪状态 BYTE 0 未知，1 准备就绪，2 故障
 	public byte bmsReadyState;
