+++ charge-pile-machine-server/src/main/java/com/huamar/charge/pile/entity/dto/resp/McCommonResp.java
@@ -26,7 +26,7 @@ public class McCommonResp extends BaseResp {
 	/**
 	 * 原消息序号
 	 */
-	private short msgNumber;
+	private int msgNumber;
 
 	/**
 	 * 消息结果
